fix(): CleanCode



# message auto-generated for no-merge-commit merge:
!234 merge clean-code-0312 into master

fix(): CleanCode

Created-by: fuyunde
Commit-by: fuyunde;chenliang267;HuRui_;cuishuang1;littlemons;suchengliu;z30066852;sqw;yjzz1007;xiaojiang;江佳瑶;fanzhongheng
Merged-by: cann-robot
Description: ## 描述
编码规范告警修改
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR不涉及Issue，可填写"NA"。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于构造对应xx测试用例、二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/hccl!234

+++ src/ops/scatter/executor/ins_v2_scatter_parallel_executor.cc
@@ -292,7 +292,7 @@ HcclResult InsV2ScatterParallelExecutor<AlgTopoMatch, InsAlgTemplate0, InsAlgTem
 
     for (u32 loopIndex = 0; loopIndex < loopTimes; loopIndex++) {
         u64 currCount = (loopIndex == loopTimes - 1) ? (dataCount_ - loopIndex * maxCountPerLoop) : maxCountPerLoop;
-        u64 dataCountPerLoopAixs0 = dataSplitSize.at(0) * currCount;
+        u64 dataCountPerLoopAixs0 = static_cast<u64>(dataSplitSize.at(0) * currCount);
         u64 dataCountPerLoopAixs1 = currCount - dataCountPerLoopAixs0;
 
         u64 dataOffset0 = loopIndex * maxCountPerLoop * dataTypeSize_;
